+++ src/web/portal/mod.rs
@@ -24,25 +24,25 @@ pub fn create_portal_routes(state: AppState) -> Router<AppState> {
     // If a member ever hits /portal/admin directly, axum returns 404
     // and the user can use the navigation.
     let admin_routes = Router::new()
-        .route("/members", get(admin::members::admin_members_page))
+        .route("/members", get(admin::members::list::admin_members_page))
         .route("/members/export", get(admin::members::admin_members_export))
         .route("/members/import", get(admin::members::admin_members_import_page))
         .route("/members/import", post(admin::members::admin_members_import))
-        .route("/members/new", get(admin::members::admin_new_member_page))
-        .route("/members/new", post(admin::members::admin_create_member))
-        .route("/members/:id", get(admin::members::admin_member_detail_page))
-        .route("/members/:id/update", post(admin::members::admin_update_member))
-        .route("/members/:id/activate", post(admin::members::admin_activate_member))
-        .route("/members/:id/suspend", post(admin::members::admin_suspend_member))
-        .route("/members/:id/extend-dues", post(admin::members::admin_extend_dues))
-        .route("/members/:id/set-dues", post(admin::members::admin_set_dues))
-        .route("/members/:id/expire-now", post(admin::members::admin_expire_now))
-        .route("/members/:id/payments", get(admin::members::admin_member_payments))
-        .route("/members/:id/record-payment", get(admin::members::admin_record_payment_page))
-        .route("/members/:id/record-payment", post(admin::members::admin_record_payment_submit))
+        .route("/members/new", get(admin::members::create::admin_new_member_page))
+        .route("/members/new", post(admin::members::create::admin_create_member))
+        .route("/members/:id", get(admin::members::detail::admin_member_detail_page))
+        .route("/members/:id/update", post(admin::members::detail::admin_update_member))
+        .route("/members/:id/activate", post(admin::members::status::admin_activate_member))
+        .route("/members/:id/suspend", post(admin::members::status::admin_suspend_member))
+        .route("/members/:id/extend-dues", post(admin::members::dues::admin_extend_dues))
+        .route("/members/:id/set-dues", post(admin::members::dues::admin_set_dues))
+        .route("/members/:id/expire-now", post(admin::members::status::admin_expire_now))
+        .route("/members/:id/payments", get(admin::members::dues::admin_member_payments))
+        .route("/members/:id/record-payment", get(admin::members::payments::admin_record_payment_page))
+        .route("/members/:id/record-payment", post(admin::members::payments::admin_record_payment_submit))
         .route("/payments/:id/refund", post(admin::payments::admin_refund_payment))
-        .route("/members/:id/resend-verification", post(admin::members::admin_resend_verification))
-        .route("/members/:id/discord-id", post(admin::members::admin_update_discord_id))
+        .route("/members/:id/resend-verification", post(admin::members::verification::admin_resend_verification))
+        .route("/members/:id/discord-id", post(admin::members::discord::admin_update_discord_id))
         // Events
         .route("/events", get(admin::events::admin_events_page))
         .route("/events/new", get(admin::events::admin_new_event_page))
